Tyler disk buffer

+++ src/collector/collector_forwarder.c
@@ -24,6 +24,8 @@
  *
  */
 
+#include "config.h"
+
 #include <string.h>
 #include <stdlib.h>
 #include <stdio.h>
@@ -31,6 +33,7 @@
 #include <assert.h>
 #include <unistd.h>
 #include <sys/timerfd.h>
+#include <amqp_tcp_socket.h>
 
 #include "util.h"
 #include "logger.h"
@@ -39,6 +42,8 @@
 
 #define BUF_BATCH_SIZE (100 * 1024 * 1024)
 #define MIN_SEND_AMOUNT (1 * 1024 * 1024)
+#define AMPQ_BYTES_FROM(x) (amqp_bytes_t){.len=sizeof(x),.bytes=&x}
+#define AMQP_FRAME_MAX 131072
 
 static inline void free_encoded_result(openli_encoded_result_t *res) {
     if (res->liid) {
@@ -50,8 +55,17 @@ static inline void free_encoded_result(openli_encoded_result_t *res) {
     }
 
     if (res->msgbody) {
+
         if (res->msgbody->encoded) {
+#ifdef HAVE_BER_ENCODING
+            if (!res->child) {
+                //if child exists, then msgbody->encoded is owned by child
+                //dont free so it can be reused
+                free(res->msgbody->encoded);
+            }
+#else
             free(res->msgbody->encoded);
+#endif
         }
         free(res->msgbody);
     }
@@ -59,6 +73,12 @@ static inline void free_encoded_result(openli_encoded_result_t *res) {
     if (res->origreq) {
         free_published_message(res->origreq);
     }
+
+#ifdef HAVE_BER_ENCODING
+    if (res->child){
+        wandder_free_child(res->child);
+    }
+#endif
 }
 
 static int add_new_destination(forwarding_thread_data_t *fwd,
@@ -71,6 +91,8 @@ static int add_new_destination(forwarding_thread_data_t *fwd,
     JLG(jval, fwd->destinations_by_id, msg->data.med.mediatorid);
 
     if (jval == NULL) {
+        char stringspace[32];
+
         newdest = (export_dest_t *)calloc(1, sizeof(export_dest_t));
 
         newdest->fd = -1;
@@ -85,6 +107,13 @@ static int add_new_destination(forwarding_thread_data_t *fwd,
         newdest->ssl = NULL;
         newdest->ssllasterror = 0;
         newdest->waitingforhandshake = 0;
+
+        if (fwd->ampq_conn) {
+            snprintf(stringspace, 32, "ID%d", newdest->mediatorid);
+
+            newdest->rmq_queueid.len = strlen(stringspace);
+            newdest->rmq_queueid.bytes = (void *)(strdup(stringspace));
+        }
 
         init_export_buffer(&(newdest->buffer));
 
@@ -108,6 +137,7 @@ static int add_new_destination(forwarding_thread_data_t *fwd,
                     strcmp(found->portstr, msg->data.med.portstr) != 0) {
                 /* Mediator has changed IP or port */
                 logger(LOG_INFO, "OpenLI: mediator %u has changed location from %s:%s to %s:%s",
+                        found->mediatorid,
                         found->ipstr, found->portstr, msg->data.med.ipstr,
                         msg->data.med.portstr);
                 free(found->ipstr);
@@ -120,6 +150,9 @@ static int add_new_destination(forwarding_thread_data_t *fwd,
                     close(found->fd);
                     found->fd = -1;
                 }
+            } else {
+                free(msg->data.med.ipstr);
+                free(msg->data.med.portstr);
             }
             found->awaitingconfirm = 0;
             found->halted = 0;
@@ -150,7 +183,9 @@ static int add_new_destination(forwarding_thread_data_t *fwd,
 static inline void disconnect_mediator(forwarding_thread_data_t *fwd,
         export_dest_t *med) {
 
-    close(med->fd);
+    if (med->fd != -1) {
+        close(med->fd);
+    }
     med->fd = -1;
 
     if (med->logallowed) {
@@ -188,7 +223,9 @@ static void remove_destination(forwarding_thread_data_t *fwd,
     if (med->portstr) {
         free(med->portstr);
     }
-
+    if (med->rmq_queueid.bytes) {
+        free(med->rmq_queueid.bytes);
+    }
 
     free(med);
 }
@@ -269,6 +306,11 @@ static void flag_all_destinations(forwarding_thread_data_t *fwd) {
         med->awaitingconfirm = 1;
     }
     fwd->awaitingconfirm = 1;
+
+    if (fwd->flagtimerfd != -1) {
+        close(fwd->flagtimerfd);
+        fwd->flagtimerfd = -1;
+    }
 }
 
 static int handle_ctrl_message(forwarding_thread_data_t *fwd,
@@ -286,7 +328,6 @@ static int handle_ctrl_message(forwarding_thread_data_t *fwd,
         free(msg->data.cept.liid);
         free(msg->data.cept.authcc);
         free(msg->data.cept.delivcc);
-        free(msg);
     } else if (msg->type == OPENLI_EXPORT_MEDIATOR) {
         return add_new_destination(fwd, msg);
     } else if (msg->type == OPENLI_EXPORT_DROP_SINGLE_MEDIATOR) {
@@ -297,6 +338,7 @@ static int handle_ctrl_message(forwarding_thread_data_t *fwd,
         if (jval == NULL) {
             logger(LOG_DEBUG, "asked to remove mediator %d but cannot find it?",
                     msg->data.med.mediatorid);
+            free(msg);
             return 1;
         }
         med = (export_dest_t *)(*jval);
@@ -311,21 +353,7 @@ static int handle_ctrl_message(forwarding_thread_data_t *fwd,
         logger(LOG_DEBUG, "causing all mediators to reconnect");
         disconnect_all_destinations(fwd);
     }
-
-    return 1;
-}
-
-static inline int enqueue_raw(forwarding_thread_data_t *fwd,
-        export_dest_t *med, openli_encoded_result_t *res) {
-
-    if (append_message_to_buffer(&(med->buffer), res, 0) == 0) {
-        logger(LOG_INFO,
-                "OpenLI: forced to drop mediator %s:%s because we cannot buffer any more records for it -- please investigate asap!",
-                med->ipstr, med->portstr);
-        remove_destination(fwd, med);
-        return 1;
-    }
-
+    free(msg);
     return 1;
 }
 
@@ -334,7 +362,6 @@ static inline int enqueue_result(forwarding_thread_data_t *fwd,
 
     PWord_t jval;
     int_reorderer_t *reord;
-    openli_encoded_result_t *rescopy;
     Pvoid_t *reorderer;
     stored_result_t *stored, *tmp;
 
@@ -349,9 +376,9 @@ static inline int enqueue_result(forwarding_thread_data_t *fwd,
 
 
     /* reordering of results if required for each LIID/CIN */
-    JSLG(jval, *reorderer, res->cinstr);
+    JSLG(jval, *reorderer, (unsigned char *)res->cinstr);
     if (jval == NULL) {
-        JSLI(jval, *reorderer, res->cinstr);
+        JSLI(jval, *reorderer, (unsigned char *)res->cinstr);
 
         if (jval == NULL) {
             logger(LOG_INFO,
@@ -385,8 +412,8 @@ static inline int enqueue_result(forwarding_thread_data_t *fwd,
 
     if (append_message_to_buffer(&(med->buffer), res, 0) == 0) {
         logger(LOG_INFO,
-                "OpenLI: forced to drop mediator %s:%s because we cannot buffer any more records for it -- please investigate asap!",
-                med->ipstr, med->portstr);
+                "OpenLI: forced to drop mediator %u because we cannot buffer any more records for it -- please investigate now!",
+                med->mediatorid);
         remove_destination(fwd, med);
         return 1;
     }
@@ -402,10 +429,10 @@ static inline int enqueue_result(forwarding_thread_data_t *fwd,
 
         if (append_message_to_buffer(&(med->buffer), &(stored->res), 0) == 0) {
             logger(LOG_INFO,
-                    "OpenLI: forced to drop mediator %s:%s because we cannot buffer any more records for it -- please investigate asap!",
-                    med->ipstr, med->portstr);
+                    "OpenLI: forced to drop mediator %u because we cannot buffer any more records for it -- please investigate asap!",
+                    med->mediatorid);
             remove_destination(fwd, med);
-            break;
+            return -1;
         }
         reord->expectedseqno = stored->res.seqno + 1;
 
@@ -459,7 +486,7 @@ static int handle_encoded_result(forwarding_thread_data_t *fwd,
 
     ret = enqueue_result(fwd, med, res);
 
-    if (ret == 1) {
+    if (ret != 0) {
         free_encoded_result(res);
     }
 
@@ -545,7 +572,7 @@ static int connect_single_target(export_dest_t *dest, SSL_CTX *ctx) {
 
 static void connect_export_targets(forwarding_thread_data_t *fwd) {
 
-    export_dest_t *dest, *tmp;
+    export_dest_t *dest;
     int ind;
     PWord_t jval, jval2;
     Word_t index;
@@ -571,6 +598,22 @@ static void connect_export_targets(forwarding_thread_data_t *fwd) {
         pthread_mutex_unlock(&(fwd->sslmutex));
         if (dest->fd == -1) {
             continue;
+        }
+
+        if (fwd->ampq_conn) {
+            amqp_queue_declare_ok_t *queue_result = amqp_queue_declare(
+                    fwd->ampq_conn,
+                    1,
+                    dest->rmq_queueid,
+                    0,
+                    1,
+                    0,
+                    0,
+                    amqp_empty_table);
+
+            if (amqp_get_rpc_reply(fwd->ampq_conn).reply_type != AMQP_RESPONSE_NORMAL ) {
+                logger(LOG_INFO, "OpenLI: Failed to declare queue");
+            }
         }
 
         JLI(jval2, fwd->destinations_by_fd, dest->fd);
@@ -608,6 +651,33 @@ static void connect_export_targets(forwarding_thread_data_t *fwd) {
         fwd->topoll[ind].revents = 0;
     }
 
+}
+
+static int drain_incoming_etsi(forwarding_thread_data_t *fwd) {
+
+    int x, encoders_over = 0;
+    openli_encoded_result_t res;
+
+    do {
+        x = zmq_recv(fwd->zmq_pullressock, &res, sizeof(res),
+                ZMQ_DONTWAIT);
+        if (x < 0 && errno != EAGAIN) {
+            return -1;
+        }
+
+        if (x < 0) {
+            continue;
+        }
+
+        if (res.liid == NULL && res.destid == 0) {
+            logger(LOG_INFO, "encoder %d has ceased encoding", encoders_over);
+            encoders_over ++;
+        }
+
+        free_encoded_result(&res);
+    } while (encoders_over < fwd->encoders);
+
+    return 1;
 }
 
 static int receive_incoming_etsi(forwarding_thread_data_t *fwd) {
@@ -665,13 +735,86 @@ static int process_control_message(forwarding_thread_data_t *fwd) {
     return 1;
 }
 
+static void rmq_write_buffered(forwarding_thread_data_t *fwd) {
+
+    export_dest_t *dest;
+    PWord_t jval;
+    uint64_t availsend = 0;
+    Word_t index = 0;
+
+    JLF(jval, fwd->destinations_by_id, index);
+    while (jval) {
+        dest = (export_dest_t *)(*jval);
+        JLN(jval, fwd->destinations_by_id, index);
+
+        if (dest->fd != -1 && fwd->forcesend_rmq &&
+                !dest->waitingforhandshake) {
+            /* XXX Warning: will block */
+            if (transmit_heartbeat(dest->fd, dest->ssl) < 0) {
+                logger(LOG_INFO,
+                        "OpenLI: failed to send heartbeat to mediator %s:%s",
+                        dest->ipstr, dest->portstr);
+                disconnect_mediator(fwd, dest);
+            }
+        }
+
+        availsend = get_buffered_amount(&(dest->buffer));
+        if (availsend == 0) {
+            continue;
+        }
+
+        if (availsend < MIN_SEND_AMOUNT && fwd->forcesend_rmq == 0) {
+            continue;
+        }
+
+        if (transmit_buffered_records_RMQ(&(dest->buffer), 
+                fwd->ampq_conn,
+                1,
+                amqp_cstring_bytes(""),
+                dest->rmq_queueid,
+                BUF_BATCH_SIZE) < 0 ) {
+            logger(LOG_INFO, "OpenLI: Error Publishing to RMQ");
+        }
+    }
+}
+
+static void complete_ssl_handshake(forwarding_thread_data_t *fwd,
+        export_dest_t *dest) {
+
+    //either keep running handshake or fail when error
+    int ret = SSL_connect(dest->ssl);
+
+    if (ret <= 0) {
+        ret = SSL_get_error(dest->ssl, ret);
+        if(ret == SSL_ERROR_WANT_READ || ret == SSL_ERROR_WANT_WRITE){
+            //keep trying
+            return;
+        }
+        else {
+            //fail out
+            if (dest->ssllasterror == 0) {
+                logger(LOG_INFO,
+                        "OpenLI: error in continuing SSL handshake with mediator: %s:%s",
+                        dest->ipstr, dest->portstr);
+            }
+            dest->waitingforhandshake = 0;
+            dest->ssllasterror = 1;
+            disconnect_mediator(fwd, dest);
+        }
+    } else {
+        logger(LOG_DEBUG, "OpenLI: SSL Handshake from mediator accepted");
+        dest->waitingforhandshake = 0;
+        dest->ssllasterror = 0;
+    }
+}
+
 static inline int forwarder_main_loop(forwarding_thread_data_t *fwd) {
     int topollc, x, i;
 
     /* Add the mediator confirmation timer to our poll item list, if
      * required.
      */
-    if (fwd->awaitingconfirm) {
+    if (fwd->awaitingconfirm && fwd->flagtimerfd != -1) {
         fwd->topoll[fwd->nextpoll].socket = NULL;
         fwd->topoll[fwd->nextpoll].fd = fwd->flagtimerfd;
         fwd->topoll[fwd->nextpoll].events = ZMQ_POLLIN;
@@ -705,6 +848,7 @@ static inline int forwarder_main_loop(forwarding_thread_data_t *fwd) {
             fwd->forcesend[i] = 1;
         }
 
+        fwd->forcesend_rmq = 1;
         its.it_interval.tv_sec = 0;
         its.it_interval.tv_nsec = 0;
         its.it_value.tv_sec = 1;
@@ -721,7 +865,7 @@ static inline int forwarder_main_loop(forwarding_thread_data_t *fwd) {
         fwd->topoll[1].revents = 0;
     }
 
-    if (fwd->awaitingconfirm) {
+    if (fwd->awaitingconfirm && fwd->flagtimerfd != -1) {
         if (fwd->topoll[fwd->nextpoll].revents & ZMQ_POLLIN) {
             purge_unconfirmed_mediators(fwd);
             fwd->awaitingconfirm = 0;
@@ -729,6 +873,15 @@ static inline int forwarder_main_loop(forwarding_thread_data_t *fwd) {
             fwd->flagtimerfd = -1;
         }
     }
+
+    if (fwd->ampq_conn) {
+        /* Loop over all destinations and see if they have anything to
+         * write to their queue.
+         */
+        rmq_write_buffered(fwd);
+        fwd->forcesend_rmq = 0;
+    }
+
 
     for (i = 3; i < fwd->nextpoll; i++) {
         export_dest_t *dest;
@@ -755,32 +908,7 @@ static inline int forwarder_main_loop(forwarding_thread_data_t *fwd) {
         dest = (export_dest_t *)(*jval);
 
         if (dest->waitingforhandshake){
-
-            int ret = SSL_connect(dest->ssl); //either keep running handshake or fail when error 
-
-            if (ret <= 0){
-                ret = SSL_get_error(dest->ssl, ret);
-                if(ret == SSL_ERROR_WANT_READ || ret == SSL_ERROR_WANT_WRITE){
-                    //keep trying
-                }
-                else {
-                    //fail out
-                    if (dest->ssllasterror == 0) {
-                        logger(LOG_INFO,
-                                "OpenLI: error in continuing SSL handshake with mediator: %s:%s",
-                                dest->ipstr, dest->portstr);
-                    }
-                    dest->waitingforhandshake = 0;
-                    dest->ssllasterror = 1;
-                    disconnect_mediator(fwd, dest);
-                    continue;
-                }
-            }
-            else {
-                logger(LOG_DEBUG, "OpenLI: SSL Handshake from mediator accepted");
-                dest->waitingforhandshake = 0;
-                dest->ssllasterror = 0;
-            }
+            complete_ssl_handshake(fwd, dest);
             continue;
         }
 
@@ -808,14 +936,13 @@ static inline int forwarder_main_loop(forwarding_thread_data_t *fwd) {
             dest->logallowed = 1;
         }
         fwd->forcesend[i] = 0;
-
     }
     return 1;
 }
 
 static void forwarder_main(forwarding_thread_data_t *fwd) {
 
-    int halted = 0, x, i;
+    int x;
     struct itimerspec its;
 
     fwd->destinations_by_id = NULL;
@@ -842,7 +969,7 @@ static void forwarder_main(forwarding_thread_data_t *fwd) {
 
     fwd->topoll = (zmq_pollitem_t *)calloc(10, sizeof(zmq_pollitem_t));
     fwd->forcesend = (uint8_t *)calloc(10, sizeof(uint8_t));
-
+    fwd->forcesend_rmq = 0;
     fwd->pollsize = 10;
     fwd->nextpoll = 3;
 
@@ -862,6 +989,11 @@ static void forwarder_main(forwarding_thread_data_t *fwd) {
 
     remove_reorderers(fwd, NULL, &(fwd->intreorderer_cc));
     remove_reorderers(fwd, NULL, &(fwd->intreorderer_iri));
+
+    if (x == 0) {
+        drain_incoming_etsi(fwd);
+    }
+
     free(fwd->topoll);
     fwd->topoll = NULL;
     free(fwd->forcesend);
@@ -915,6 +1047,41 @@ void *start_forwarding_thread(void *data) {
         goto haltforwarder;
     }
 
+    if (fwd->RMQ_conf.enabled) {
+        if ( fwd->RMQ_conf.name && fwd->RMQ_conf.pass ) {
+            fwd->ampq_conn = amqp_new_connection();
+            fwd->ampq_sock = amqp_tcp_socket_new(fwd->ampq_conn);
+
+            //TODO RMQ instance will always be on localhost? (for collector)
+            if (amqp_socket_open(fwd->ampq_sock, "localhost", 5672 )){
+                logger(LOG_INFO, 
+                        "OpenLI: RMQ forwarding thread %d failed to open amqp socket",
+                        fwd->forwardid);
+                goto haltforwarder;
+            }
+
+            /* login using PLAIN, must specify username and password */
+            if ( (amqp_login(fwd->ampq_conn, "OpenLI", 0, AMQP_FRAME_MAX,0,
+                            AMQP_SASL_METHOD_PLAIN, fwd->RMQ_conf.name,
+                            fwd->RMQ_conf.pass)
+                    ).reply_type != AMQP_RESPONSE_NORMAL ) {
+                logger(LOG_ERR, "OpenLI: RMQ Failed to login to broker using PLAIN auth");
+                goto haltforwarder;
+            }
+
+            amqp_channel_open_ok_t *r = amqp_channel_open(fwd->ampq_conn, 1);
+
+            if ( (amqp_get_rpc_reply(fwd->ampq_conn).reply_type) != AMQP_RESPONSE_NORMAL ) {
+                logger(LOG_ERR, "OpenLI: Failed to open channel");
+                goto haltforwarder;
+            }
+            logger(LOG_INFO, "OpenLI: Connected to RMQ instance");
+        } else {
+            logger(LOG_INFO, "OpenLI: Incomplete RMQ login information supplied");
+            goto haltforwarder;
+        }
+    }
+
     forwarder_main(fwd);
 
     do {
@@ -941,6 +1108,9 @@ void *start_forwarding_thread(void *data) {
     } while (x > 0);
 
 haltforwarder:
+    if (fwd->ampq_conn){
+        amqp_destroy_connection(fwd->ampq_conn);
+    }
     zmq_close(fwd->zmq_ctrlsock);
     remove_all_destinations(fwd);
     logger(LOG_DEBUG, "OpenLI: halting forwarding thread %d",
